--- services/frontend/src/components/Calendar/Month/Loose/Status/constants.ts
@@ -1,1 +1,0 @@
-export const MAX_LENGTH = 3;
