+++ services/frontend/src/components/Calendar/Month/index.ts
@@ -1,2 +1,1 @@
-export { default as CalendarMonthLoose } from './Loose';
 export { default as CalendarMonthTight } from './Tight';
