+++ services/frontend/src/features/booking/detail/components/Footer/i.json
