+++ services/frontend/src/features/booking/detail/components/Footer/index.tsx
