+++ services/frontend/src/features/booking/detail/components/ServiceDetail/components/InfoBlock/index.tsx
